requirements.txt
numpy
pandas
openai
python-dotenv
requests
tqdm
